Api development

+++ backend_scraper/test_app.py
@@ -29,13 +29,41 @@ async def update_manga_list(manga_list: MangaList):
     # Process the manga_list as required
     # For now, let's just return a confirmation message
 
-    with open('manga_data.json', 'w') as file:
-        file.write(manga_list.json())
+    # with open('manga_data.json', 'w') as file:
+    #     file.write(manga_list.json())
+    output_list, error_list = scrape_record(manga_list)
+    bulk_insert_record(output_list)
+    delete_record(manga_list)
+    ms_db = MangaScraperDB()
+    response = ms_db.get_frontend_data()
+    print(f"response: {response}")
     return {
         "message": "Successfully confirmed", 
         "url": "http://127.0.0.1:8000/", 
-        "confirmation": manga_list
+        "confirmation": response
     }
+
+## TODO: Handling duplicate records that are inserted to the database. Current method just adds new row with new ID even if data is the same
+def bulk_insert_record(output_list: List[Dict[str, Any]]):
+    """
+    Bulk insert records then close at the end
+
+    Args:
+        output_list (List[Dict[str, Any]]): _description_
+    """
+    ms_db = MangaScraperDB()
+    for item in output_list:
+        manga_name = item["manga_name"]
+        print(manga_name)
+        manga_id = ms_db.insert_manga(manga_name = manga_name)
+
+        website_id = ms_db.get_website_id(item["website_url"])
+        manga_path = item["manga_path"]
+        manga_path_id = ms_db.insert_manga_path(manga_id = manga_id, website_id = website_id, manga_path = manga_path)
+        manga_chapter_url_id = ms_db.insert_manga_chapter_url_store(record = item, manga_id = manga_id, website_id = website_id, manga_path_id = manga_path_id)
+        manga_thumbnail_id = ms_db.insert_manga_thumbnail(manga_id, website_id, manga_path_id, thumbnail_url=item["manga_thumbnail_url"])
+        print(manga_id, website_id, manga_path, manga_path_id, manga_chapter_url_id, manga_thumbnail_id)
+    ms_db.close_connection()
 
 def scrape_record(manga_list):
     """
@@ -78,18 +106,9 @@ def get_new_record(manga_list):
     # From the provided manga list
     # Need to identify all the new ones can create new data structure for them
     # These are item's with a "new_" prefix in the ID
-    new_list = [item for item in manga_list if "new_" in item.id]
+    
+    new_list = [item for item in manga_list.manga_records if "new_" in item.id]
     return new_list
-
-def insert_records(output_list):
-    """
-    Method to call the mangaScraperDB class to perform DB operations
-
-    Args:
-        output_list (_type_): _description_
-    """
-    pass
-
 
 @app.get("/get_data", response_model=List[Dict[str, Any]])
 async def get_data() -> List[Dict[str, Any]]:
@@ -103,7 +122,32 @@ async def get_data() -> List[Dict[str, Any]]:
     manga_list = ms_db.get_frontend_data()
     return manga_list
 
+@app.get("/get_bookmarks_data", response_model=List[Dict[str, Any]])
+async def get_data() -> List[Dict[str, Any]]:
+    """
+    Endpoint to retrieve manga data for the frontend bookmarks component
 
+    Returns:
+        List[Dict[str, Any]]: A list of dictionaries containing manga data.
+    """
+    ms_db = MangaScraperDB()
+    manga_list = ms_db.get_bookmarks_data()
+    return manga_list
+
+@app.get("/get_supported_websites", response_model=List[Dict[str, Any]])
+async def get_data() -> List[Dict[str, Any]]:
+    """
+    Endpoint to retrieve manga data for the frontend supported websites component within bookmarks
+
+    Returns:
+        List[Dict[str, Any]]: A list of dictionaries containing manga data.
+    """
+    ms_db = MangaScraperDB()
+    manga_list = ms_db.get_supported_websites()
+    print(manga_list)
+    return manga_list
+
+## TODO: Migrate delete record to the ms_db class
 def delete_record(manga_list: List[MangaRecord]) -> List[Dict[str, str]]:
     """
     Deletes records from the database for each manga in the provided list that is marked with the status 'Delete'.
@@ -120,18 +164,18 @@ def delete_record(manga_list: List[MangaRecord]) -> List[Dict[str, str]]:
                               to be deleted. Each dictionary includes the 'id' of the manga and the 'error' message.
     """
 
-    ms_db = MangaScraperDB()
-    delete_list = [item for item in manga_list if item.status.lower() == "delete"]
-
+    delete_list = [item for item in manga_list.manga_records if item.status.lower() == "delete"]
     error_list = []
-    for item in delete_list:
-        try:
-            with ms_db.conn.cursor() as cur:
-                cur.execute("CALL delete_manga_record(%s)", (item.id,))
-                ms_db.conn.commit()
-        except Exception as e:
-            error_list.append({"id": item.id, "error": str(e)})
-            ms_db.conn.rollback()
+    if len(delete_list) >0:
+        ms_db = MangaScraperDB()
+        for item in delete_list:
+            try:
+                with ms_db.conn.cursor() as cur:
+                    cur.execute("CALL delete_manga_record(%s)", (item.id,))
+                    ms_db.conn.commit()
+            except Exception as e:
+                error_list.append({"id": item.id, "error": str(e)})
+                ms_db.conn.rollback()
 
-    ms_db.close_connection()
+        ms_db.close_connection()
     return error_list
